move `import_yaml` to util.py
centralize helper function
