+++ xpdacq/utils.py
@@ -94,3 +94,37 @@ def export_userScriptEtc():
         _graceful_exit('Did you accidentally change write privilege to {}'.format(glbl.home))
         print('Please check your setting and try `export_userScriptEtc()` again at command prompt')
         return
+
+def import_yaml():
+    '''
+    import user pre-defined files from ~/xpdUser/Import
+
+    Files can be compreesed or .yml, once imported, bt.list() should show updated acquire object list
+    '''
+    src_dir = glbl.import_dir
+    dst_dir = glbl.yaml_dir
+    f_list = os.listdir(src_dir)
+    if len(f_list) == 0:
+        print('INFO: There is no pre-defined user objects in {}'.format(src_dir))
+        return 
+    # two possibilites: .yml or compressed files; shutil should handle all compressed cases
+    moved_f_list = []
+    for f in f_list:
+        full_path = os.path.join(src_dir, f)
+        (root, ext) = os.path.splitext(f)
+        if ext == '.yml':
+            shutil.copy(full_path, dst_dir)
+            moved_f_list.append(f)
+            # FIXME - do we want user confirmation?
+            os.remove(full_path)
+        else:
+            try:
+                shutil.unpack_archive(full_path, dst_dir)
+                moved_f_list.append(f)
+                # FIXME - do we want user confirmation?
+                os.remove(full_path)
+            except ReadError:
+                print('Unrecongnized file type {} is found inside {}'.format(f, src_dir))
+                pass
+    return moved_f_list
+
